address feedback: serde trait
